tests/wpt/tests/html/canvas/element/the-canvas-state/2d.state.saverestore.shadowBlur.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.state.saverestore.shadowBlur</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.state.saverestore.shadowBlur</h1>
<p class="desc">save()/restore() works for shadowBlur</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("save()/restore() works for shadowBlur");
_addTest(function(canvas, ctx) {

  // Test that restore() undoes any modifications
  var old = ctx.shadowBlur;
  ctx.save();
  ctx.shadowBlur = 5;
  ctx.restore();
  _assertSame(ctx.shadowBlur, old, "ctx.shadowBlur", "old");

  // Also test that save() doesn't modify the values
  ctx.shadowBlur = 5;
  old = ctx.shadowBlur;
      // we're not interested in failures caused by get(set(x)) != x (e.g.
      // from rounding), so compare against 'old' instead of against 5
  ctx.save();
  _assertSame(ctx.shadowBlur, old, "ctx.shadowBlur", "old");
  ctx.restore();

});
</script>
